pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rlinf"
version = "0.2.0"
authors = [{ name = "RLinf Contributors" }]
description = "Reinforcement Learning Infrastructure for Embodied and Agentic AI"
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.10,<=3.11.14"
license = {file = "LICENSE"}
keywords = [
    "reinforcement-learning",
    "embodied-intelligence",
    "large-language-models",
]
classifiers = [
    #   2 - Pre-Alpha
    #   3 - Alpha
    #   4 - Beta
    #   5 - Production/Stable
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: GPU :: NVIDIA CUDA :: 12 :: 12.4",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]


dependencies = [
    # Core System
    # These are the dependencies of the core scheduler
    "ray[default]>=2.47.0",
    "torch>=2.5.0,<=2.9.0",
    "hydra-core",
    "numpy",

    # Data processing
    "datasets",
    "torchdata",
    "scipy",

    # Utilities
    "accelerate",
    "debugpy",
    "einops",
    "nvitop",
    "pybind11",
    "setuptools",
    "ninja",
    "pytest",
    "gsutil",
    "ruff",
    "pre-commit",
    "liger_kernel; sys_platform == 'linux' and platform_machine == 'x86_64'",
    "uv",
    "pip",
    "huggingface_hub",
    "icmplib",
    "pyrealsense2; sys_platform == 'linux' and platform_machine == 'x86_64'",

    # Logging
    "wandb<0.25.1",
    "swanlab>=0.6.11",
    "tensorboard",
]

[project.urls]
Homepage = "https://github.com/RLinf/RLinf"
Repository = "https://github.com/RLinf/RLinf"
Documentation = "https://rlinf.readthedocs.io/"
"Bug Report" = "https://github.com/RLinf/RLinf/issues"

[project.optional-dependencies]
agentic-sglang = [
    "sglang[all]==0.4.6.post5",
    "torch-memory-saver",
    "numpy==2.2",
    "transformers==4.51.1",
    # uvloop is not compatible with simulators, put it here
    "uvloop==0.21.0",
    "hydra-core==1.4.0.dev1",
    "omegaconf==2.4.0.dev4",
    "word2number",
    "regex",
    "peft==0.11.1",
    "rlinf_latex2sympy2",
    "mcp",
]
agentic-vllm = [
    "vllm==0.8.5",
    "numpy==2.2",
    "transformers==4.51.1",
    # uvloop is not compatible with simulators, put it here
    "uvloop==0.21.0",
    "hydra-core==1.4.0.dev1",
    "omegaconf==2.4.0.dev4",
    "word2number",
    "regex",
    "peft==0.11.1",
    "rlinf_latex2sympy2",
    "mcp",
]
embodied = [
    "transformers<=4.57.6",
    "peft",
    "timm",
    "imageio[ffmpeg]",
    "gymnasium",
    "gym",
]
a2d = [
    "grpcio>=1.62.0",
    "protobuf>=6.31.1",
]
franka = [
    "pyyaml",
    "rospkg",
    "empy==3.3.4",
    "gymnasium==0.29.1",
    "pyrealsense2",
    "pyquaternion",
    "opencv-python",
    "defusedxml",
    "pyspacemouse==1.1.5", 
    "imageio[ffmpeg]",
    "pynput"
]
xsquare_turtle2 = [
    "pyyaml",
    "rospkg",
    "toppra",
    "numpy==1.26.4",
    "empy==3.3.4",
    "gymnasium==0.29.1",
    "pyrealsense2",
    "pyquaternion",
    "opencv-python",
    "defusedxml",
    "pyspacemouse==1.1.5", 
    "imageio[ffmpeg]",
    "pynput"
]

[tool.uv]
prerelease = "allow"
conflicts = [
    [
      { extra = "agentic-sglang" },
      { extra = "agentic-vllm" },
      { extra = "embodied" },
      { extra = "a2d" },
      { extra = "franka" },
      { extra = "xsquare_turtle2" }
    ],
]
override-dependencies = [
    "torch==2.6.0",
    "torchvision==0.21.0",
    "torchaudio==2.6.0",
    "torchcodec==0.2",
    "xgrammar==0.1.19",
    "uvloop==0.21.0",
    "tensorflow-addons==0.23.0",
    "setuptools>=69.5.1,<75.9",
    "datasets==3.6.0",
    "peft==0.11.1",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["rlinf*"]
exclude = ["tests*", "docs*", "examples*", "docker*", "toolkits*"]

[tool.setuptools.package-data]
rlinf = ["**/*.yaml", "**/*.yml"]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
isort = {known-first-party = ["rlinf"]}
select = ["C", "E", "F", "I", "W", "CPY001", "RUF013", "UP006", "PERF102", "PLC1802", "PLC0208", "D", "RUF002", "DOC102", "DOC202", "DOC403", "DOC502"]
ignore = [
    "C901", # "complex-structure"
    "E501", # "line-too-long"
    "E741", # "ambiguous-variable-name"
    "D100", # "Missing docstring in public module"
    "D104", # "Missing docstring in public package"
    "D203", # "incorrect-blank-line-before-class", conflict with D211,
    "D213", # "multi-line-summary-second-line", conflict with D212
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
# Only enable docstring check for the scheduler module for now.
"!rlinf/scheduler/**.py" = ["D", "RUF002"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"